+++ Reqnroll/Assist/Service.cs
@@ -23,30 +23,6 @@ namespace Reqnroll.Assist
             ValueRetrievers = new ReqnrollDefaultValueRetrieverList();
         }
 
-        [Obsolete("Use ValueComparers.Register")]
-        public void RegisterValueComparer(IValueComparer valueComparer)
-        {
-            ValueComparers.Register(valueComparer);
-        }
-
-        [Obsolete("Use ValueComparers.Unregister")]
-        public void UnregisterValueComparer(IValueComparer valueComparer)
-        {
-            ValueComparers.Unregister(valueComparer);
-        }
-
-        [Obsolete("Use ValueRetrievers.Register")]
-        public void RegisterValueRetriever(IValueRetriever valueRetriever)
-        {
-            ValueRetrievers.Register(valueRetriever);
-        }
-
-        [Obsolete("Use ValueRetrievers.Unregister")]
-        public void UnregisterValueRetriever(IValueRetriever valueRetriever)
-        {
-            ValueRetrievers.Unregister(valueRetriever);
-        }
-
         public IValueRetriever? GetValueRetrieverFor(DataTableRow row, Type targetType, Type propertyType)
         {
             var keyValuePair = new KeyValuePair<string, string>(row[0], row[1]);