+++ CHANGELOG.md
@@ -34,7 +34,14 @@
 * Fixed source link and deterministic compilation for Reqnroll.CustomPlugin package (#719)
 * Fix: Rule Tags are now properly generated as Test Categories (along with Scenario Tags) (#731)(
 
-*Contributors of this release (in alphabetical order):* @304NotModified, @algirdasN, @clrudolphi, @DrEsteban, @loraderon, @obligaron
+## Deprecations:
+
+* The synchronous test runner API (`ISyncTestRunner`) has been removed. Please use `ITestRunner` instead.
+* The synchronous code invocation API (`IBindingInvoker`) has been deprecated. Please use `IAsyncBindingInvoker` instead.
+* Removed obsolete property `ScenarioInfo.ScenarioAndFeatureTags`. Please use `ScenarioInfo.CombinedTags` instead.
+* Removed obsolete methods on `Reqnroll.Assist.Service` class: `RegisterValueComparer`, `UnregisterValueComparer`, `RegisterValueRetriever`, `UnregisterValueRetriever` use `ValueComparers.Register`, `ValueComparers.Unregister`, `ValueRetrievers.Register`, `ValueRetrievers.Unregister` instead.
+
+*Contributors of this release (in alphabetical order):* @304NotModified, @algirdasN, @clrudolphi, @DrEsteban, @gasparnagy, @loraderon, @obligaron
 
 # v2.4.1 - 2025-04-29